RefWatcher in BaseController via interface
