+++ common/src/main/java/com/alefimenko/iuttimetable/common/IUTRefWatcher.kt
@@ -1,0 +1,7 @@
+package com.alefimenko.iuttimetable.common
+
+import com.squareup.leakcanary.RefWatcher
+
+interface IUTRefWatcher {
+    var refWatcher: RefWatcher?
+}
