+++ coreui/src/main/java/com/alefimenko/iuttimetable/base/BaseController.kt
@@ -7,8 +7,12 @@ import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.bluelinelabs.conductor.archlifecycle.LifecycleController
+import com.alefimenko.iuttimetable.common.IUTRefWatcher
 import com.alefimenko.iuttimetable.createBinder
+import com.alefimenko.iuttimetable.extension.requireActivity
+import com.bluelinelabs.conductor.ControllerChangeHandler
+import com.bluelinelabs.conductor.ControllerChangeType
+import com.bluelinelabs.conductor.archlifecycle.LifecycleController
 import io.reactivex.ObservableSource
 import io.reactivex.functions.Consumer
 import io.reactivex.subjects.PublishSubject
@@ -23,6 +27,8 @@ abstract class BaseController<Event, ViewModel>(
     Consumer<ViewModel>,
     ObservableSource<Event> by events,
     ComponentCallbacks {
+
+    private var hasExited: Boolean = false
 
     open var layoutRes: Int = 0
 
@@ -65,6 +71,20 @@ abstract class BaseController<Event, ViewModel>(
         bind.resetViews()
         mainHandler.removeCallbacksAndMessages(null)
         super.onDestroyView(view)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        if (hasExited) (requireActivity().application as IUTRefWatcher).refWatcher?.watch(this)
+    }
+
+    override fun onChangeEnded(
+        changeHandler: ControllerChangeHandler,
+        changeType: ControllerChangeType
+    ) {
+        super.onChangeEnded(changeHandler, changeType)
+        hasExited = !changeType.isEnter
+        if (isDestroyed) (requireActivity().application as IUTRefWatcher).refWatcher?.watch(this)
     }
 
     override fun onLowMemory() = Unit
